Added unit and integration tests for interpreter and fixed a bug where no more then 1 model where returned.
