+++ test/blackbox/AsyncAPI.spec.ts
@@ -8,7 +8,7 @@ describe('AsyncAPI JSON Schema file', () => {
     const generator = new TypeScriptGenerator();
     const generatedContent = await generator.generate(inputSchema);
     expect(generatedContent).not.toBeUndefined();
-    expect(generatedContent.length).toBeGreaterThan(0);
+    expect(generatedContent.length).toBeGreaterThan(1);
   });
   test('should be generated in Java', async () => {
     const inputSchemaString = fs.readFileSync(path.resolve(__dirname, './docs/AsyncAPI_2_0_0.json'), 'utf8');
@@ -16,7 +16,7 @@ describe('AsyncAPI JSON Schema file', () => {
     const generator = new JavaGenerator();
     const generatedContent = await generator.generate(inputSchema);
     expect(generatedContent).not.toBeUndefined();
-    expect(generatedContent.length).toBeGreaterThan(0);
+    expect(generatedContent.length).toBeGreaterThan(1);
   });
   test('should be generated in JavaScript', async () => {
     const inputSchemaString = fs.readFileSync(path.resolve(__dirname, './docs/AsyncAPI_2_0_0.json'), 'utf8');
@@ -24,6 +24,6 @@ describe('AsyncAPI JSON Schema file', () => {
     const generator = new JavaScriptGenerator();
     const generatedContent = await generator.generate(inputSchema);
     expect(generatedContent).not.toBeUndefined();
-    expect(generatedContent.length).toBeGreaterThan(0);
+    expect(generatedContent.length).toBeGreaterThan(1);
   });
 });
