+++ test/interpreter/unit/InterpretAdditionalProperties.spec.ts
