+++ test/interpreter/unit/InterpretAllOf.spec.ts
