+++ test/interpreter/unit/InterpretEnum.spec.ts
