+++ test/interpreter/unit/InterpretNot.spec.ts
