+++ test/interpreter/unit/InterpretProperties.spec.ts
