+++ test/interpreter/unit/patternProperties.spec.ts
